+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.3 (work in progress)
 
 * added a runtime_init_hook(hook) method to the EsRuntimeBuilder so we can add vars to the runtime when the builder is built
+* added EsPromise::new_async which can be used to instantiate a Promise with an async resolver
 
 # 0.2.2